+++ validator/rpc/apimiddleware/structs_test.go
@@ -118,116 +118,6 @@ func TestDeleteKeystores_JSONisEqual(t *testing.T) {
 
 }
 
-func TestListRemoteKeys_JSONisEqual(t *testing.T) {
-	middlewareResponse := &ListRemoteKeysResponseJson{
-		Keystores: []*RemoteKeysListJson{
-			{
-				Pubkey:   "0x0",
-				Url:      "http://localhost:8080",
-				Readonly: true,
-			},
-		},
-	}
-
-	protoResponse := &service.ListRemoteKeysResponse{
-		Data: []*service.ListRemoteKeysResponse_Keystore{
-			{
-				Pubkey:   make([]byte, fieldparams.BLSPubkeyLength),
-				Url:      "http://localhost:8080",
-				Readonly: true,
-			},
-		},
-	}
-
-	listResp, err := areJsonPropertyNamesEqual(middlewareResponse, protoResponse)
-	require.NoError(t, err)
-	require.Equal(t, listResp, true)
-
-	resp, err := areJsonPropertyNamesEqual(middlewareResponse.Keystores[0], protoResponse.Data[0])
-	require.NoError(t, err)
-	require.Equal(t, resp, true)
-}
-
-func TestImportRemoteKeys_JSONisEqual(t *testing.T) {
-	importKeystoresRequest := &ImportRemoteKeysRequestJson{}
-
-	protoImportRequest := &service.ImportRemoteKeysRequest{
-		RemoteKeys: []*service.ImportRemoteKeysRequest_Keystore{
-			{
-				Pubkey: make([]byte, fieldparams.BLSPubkeyLength),
-				Url:    "http://localhost:8080",
-			},
-		},
-	}
-
-	requestResp, err := areJsonPropertyNamesEqual(importKeystoresRequest, protoImportRequest)
-	require.NoError(t, err)
-	require.Equal(t, requestResp, true)
-
-	importKeystoresResponse := &ImportRemoteKeysResponseJson{
-		Statuses: []*StatusJson{
-			{
-				Status:  "Error",
-				Message: "a",
-			},
-		},
-	}
-
-	protoImportKeystoresResponse := &service.ImportRemoteKeysResponse{
-		Data: []*service.ImportedRemoteKeysStatus{
-			{
-				Status:  service.ImportedRemoteKeysStatus_ERROR,
-				Message: "a",
-			},
-		},
-	}
-
-	ImportResp, err := areJsonPropertyNamesEqual(importKeystoresResponse, protoImportKeystoresResponse)
-	require.NoError(t, err)
-	require.Equal(t, ImportResp, true)
-
-	resp, err := areJsonPropertyNamesEqual(importKeystoresResponse.Statuses[0], protoImportKeystoresResponse.Data[0])
-	require.NoError(t, err)
-	require.Equal(t, resp, true)
-}
-
-func TestDeleteRemoteKeys_JSONisEqual(t *testing.T) {
-	deleteKeystoresRequest := &DeleteRemoteKeysRequestJson{}
-
-	protoDeleteRequest := &service.DeleteRemoteKeysRequest{
-		Pubkeys: [][]byte{{}},
-	}
-
-	requestResp, err := areJsonPropertyNamesEqual(deleteKeystoresRequest, protoDeleteRequest)
-	require.NoError(t, err)
-	require.Equal(t, requestResp, true)
-
-	deleteKeystoresResponse := &DeleteRemoteKeysResponseJson{
-		Statuses: []*StatusJson{
-			{
-				Status:  "Error",
-				Message: "a",
-			},
-		},
-	}
-	protoDeleteResponse := &service.DeleteRemoteKeysResponse{
-		Data: []*service.DeletedRemoteKeysStatus{
-			{
-				Status:  service.DeletedRemoteKeysStatus_ERROR,
-				Message: "a",
-			},
-		},
-	}
-
-	deleteResp, err := areJsonPropertyNamesEqual(deleteKeystoresResponse, protoDeleteResponse)
-	require.NoError(t, err)
-	require.Equal(t, deleteResp, true)
-
-	resp, err := areJsonPropertyNamesEqual(deleteKeystoresResponse.Statuses[0], protoDeleteResponse.Data[0])
-	require.NoError(t, err)
-	require.Equal(t, resp, true)
-}
-
 // note: this does not do a deep comparison of the structs
 func areJsonPropertyNamesEqual(internal, proto interface{}) (bool, error) {
 	internalJSON, err := json.Marshal(internal)
