+++ validator/rpc/standard_api_test.go
@@ -542,161 +542,6 @@ func createRandomKeystore(t testing.TB, password string) *keymanager.Keystore {
 	}
 }
 
-func TestServer_ListRemoteKeys(t *testing.T) {
-	t.Run("wallet not ready", func(t *testing.T) {
-		s := Server{}
-		_, err := s.ListKeystores(context.Background(), &empty.Empty{})
-		require.ErrorContains(t, "Prysm Wallet not initialized. Please create a new wallet.", err)
-	})
-	ctx := context.Background()
-	w := wallet.NewWalletForWeb3Signer()
-	root := make([]byte, fieldparams.RootLength)
-	root[0] = 1
-	bytevalue, err := hexutil.Decode("0x93247f2209abcacf57b75a51dafae777f9dd38bc7053d1af526f220a7489a6d3a2753e5f3e8b1cfe39b56f43611df74a")
-	require.NoError(t, err)
-	pubkeys := [][fieldparams.BLSPubkeyLength]byte{bytesutil.ToBytes48(bytevalue)}
-	config := &remoteweb3signer.SetupConfig{
-		BaseEndpoint:          "http://example.com",
-		GenesisValidatorsRoot: root,
-		ProvidedPublicKeys:    pubkeys,
-	}
-	km, err := w.InitializeKeymanager(ctx, iface.InitKeymanagerConfig{ListenForChanges: false, Web3SignerConfig: config})
-	require.NoError(t, err)
-	vs, err := client.NewValidatorService(ctx, &client.Config{
-		Wallet: w,
-		Validator: &mock.MockValidator{
-			Km: km,
-		},
-		Web3SignerConfig: config,
-	})
-	require.NoError(t, err)
-	s := &Server{
-		walletInitialized: true,
-		wallet:            w,
-		validatorService:  vs,
-	}
-	expectedKeys, err := km.FetchValidatingPublicKeys(ctx)
-	require.NoError(t, err)
-
-	t.Run("returns proper data with existing pub keystores", func(t *testing.T) {
-		resp, err := s.ListRemoteKeys(context.Background(), &empty.Empty{})
-		require.NoError(t, err)
-		for i := 0; i < len(resp.Data); i++ {
-			require.DeepEqual(t, expectedKeys[i][:], resp.Data[i].Pubkey)
-		}
-	})
-}
-
-func TestServer_ImportRemoteKeys(t *testing.T) {
-	t.Run("wallet not ready", func(t *testing.T) {
-		s := Server{}
-		_, err := s.ListKeystores(context.Background(), &empty.Empty{})
-		require.ErrorContains(t, "Prysm Wallet not initialized. Please create a new wallet.", err)
-	})
-	ctx := context.Background()
-	w := wallet.NewWalletForWeb3Signer()
-	root := make([]byte, fieldparams.RootLength)
-	root[0] = 1
-	config := &remoteweb3signer.SetupConfig{
-		BaseEndpoint:          "http://example.com",
-		GenesisValidatorsRoot: root,
-		ProvidedPublicKeys:    nil,
-	}
-	km, err := w.InitializeKeymanager(ctx, iface.InitKeymanagerConfig{ListenForChanges: false, Web3SignerConfig: config})
-	require.NoError(t, err)
-	vs, err := client.NewValidatorService(ctx, &client.Config{
-		Wallet: w,
-		Validator: &mock.MockValidator{
-			Km: km,
-		},
-		Web3SignerConfig: config,
-	})
-	require.NoError(t, err)
-	s := &Server{
-		walletInitialized: true,
-		wallet:            w,
-		validatorService:  vs,
-	}
-	bytevalue, err := hexutil.Decode("0x93247f2209abcacf57b75a51dafae777f9dd38bc7053d1af526f220a7489a6d3a2753e5f3e8b1cfe39b56f43611df74a")
-	require.NoError(t, err)
-	remoteKeys := []*ethpbservice.ImportRemoteKeysRequest_Keystore{
-		{
-			Pubkey: bytevalue,
-		},
-	}
-
-	t.Run("returns proper data with existing pub keystores", func(t *testing.T) {
-		resp, err := s.ImportRemoteKeys(context.Background(), &ethpbservice.ImportRemoteKeysRequest{
-			RemoteKeys: remoteKeys,
-		})
-		expectedStatuses := []*ethpbservice.ImportedRemoteKeysStatus{
-			{
-				Status:  ethpbservice.ImportedRemoteKeysStatus_IMPORTED,
-				Message: fmt.Sprintf("Successfully added pubkey: %v", hexutil.Encode(bytevalue)),
-			},
-		}
-		require.NoError(t, err)
-		for i := 0; i < len(resp.Data); i++ {
-			require.DeepEqual(t, expectedStatuses[i], resp.Data[i])
-		}
-	})
-}
-
-func TestServer_DeleteRemoteKeys(t *testing.T) {
-	t.Run("wallet not ready", func(t *testing.T) {
-		s := Server{}
-		_, err := s.ListKeystores(context.Background(), &empty.Empty{})
-		require.ErrorContains(t, "Prysm Wallet not initialized. Please create a new wallet.", err)
-	})
-	ctx := context.Background()
-	w := wallet.NewWalletForWeb3Signer()
-	root := make([]byte, fieldparams.RootLength)
-	root[0] = 1
-	bytevalue, err := hexutil.Decode("0x93247f2209abcacf57b75a51dafae777f9dd38bc7053d1af526f220a7489a6d3a2753e5f3e8b1cfe39b56f43611df74a")
-	require.NoError(t, err)
-	pubkeys := [][fieldparams.BLSPubkeyLength]byte{bytesutil.ToBytes48(bytevalue)}
-	config := &remoteweb3signer.SetupConfig{
-		BaseEndpoint:          "http://example.com",
-		GenesisValidatorsRoot: root,
-		ProvidedPublicKeys:    pubkeys,
-	}
-	km, err := w.InitializeKeymanager(ctx, iface.InitKeymanagerConfig{ListenForChanges: false, Web3SignerConfig: config})
-	require.NoError(t, err)
-	vs, err := client.NewValidatorService(ctx, &client.Config{
-		Wallet: w,
-		Validator: &mock.MockValidator{
-			Km: km,
-		},
-		Web3SignerConfig: config,
-	})
-	require.NoError(t, err)
-	s := &Server{
-		walletInitialized: true,
-		wallet:            w,
-		validatorService:  vs,
-	}
-
-	t.Run("returns proper data with existing pub keystores", func(t *testing.T) {
-		resp, err := s.DeleteRemoteKeys(context.Background(), &ethpbservice.DeleteRemoteKeysRequest{
-			Pubkeys: [][]byte{bytevalue},
-		})
-		expectedStatuses := []*ethpbservice.DeletedRemoteKeysStatus{
-			{
-				Status:  ethpbservice.DeletedRemoteKeysStatus_DELETED,
-				Message: fmt.Sprintf("Successfully deleted pubkey: %v", hexutil.Encode(bytevalue)),
-			},
-		}
-		require.NoError(t, err)
-		for i := 0; i < len(resp.Data); i++ {
-			require.DeepEqual(t, expectedStatuses[i], resp.Data[i])
-
-		}
-		expectedKeys, err := km.FetchValidatingPublicKeys(ctx)
-		require.NoError(t, err)
-		require.Equal(t, 0, len(expectedKeys))
-	})
-}
-
 func TestServer_ListFeeRecipientByPubkey(t *testing.T) {
 	ctx := context.Background()
 	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
